refactor(shell): extract DictController from ReaderActivity
Move the dictionary subsystem (RR12 / ADR-INKREAD-0009 D2/D3) — the corpus
handle lifecycle, the on-device→Wiktionary lookup chain, the definition card,
and the user-dictionary install/remove surface — into a focused DictController
behind a narrow Host seam (activity/docHandle/engineExecute). The Activity's
input handlers and toolbar now call dict.defineWord/defineRect/
defineSelectionText/showDictionariesDialog/close.

Behavior + threading preserved (defineWord/defineRect stay engine-thread,
caller-wrapped). ReaderActivity drops ~450 more lines (now 2703, from 3302 at
the start of this branch); DictController is 513 lines. Kotlin compiles.
